// src/app/models/consultaOcurrencia.ts
export class ConsultaOcurrencia {

    codigo: string;
    usuario: string;
    data_fato: string;
    bo: string;
    obs: string;
    Nome: string;
    nome_bairro: string;
    aisp: string;
    risp: string;
    cipm: string;
    dt:string;
    logradouro:string;
    tipo_local:string;
    ponto_referencia:string;
    situacao: string;
    nomeusuario:string;
    latitude:string;
    longitude:string;
    //fOR PRINT pdf
    Datafato:string;
    hora_fato:string;
    matriculausuario:string; 
    pim:string;
    imei:string;

    DataInicial: string='';
    DataFinal: string='';
    CodigoMunicipio: any;
    codigo_bairro: any;
}